==> 2.1_Wrangle.Rmd <==
# Setup
```{r setup}
    rm(list = ls())

    # knitr options
    knitr::opts_chunk$set(warning = FALSE, message = FALSE, echo = FALSE, fig.align = 'center',
                          fig.width = 10, fig.height = 7) 
    
    # libraries 
    library(tidyverse)
    library(dplyr)
    library(ggplot2)
    library(sp)
    library(raster)
    library(rgeos)
    library(rgdal)
    library(sf)
    library(googledrive)
    library(units)
    library(nngeo)
    library(magrittr)
    library(mgcv)
    
    # data
    ## gpkg - ESPG 4283 (GDA94)
    grid <- readRDS("./data/gpkg/2.0_Sites_4283.gpkg")
    
    dat <- read_csv("data/01_data/1.2_dat.csv") %>% 
      filter(ActivityType == "Extractive" & SiteType == "Boat" & (str_detect(Site, "BR")))
    
    # dat <- read_csv("data/01_data/1.2_dat.csv") %>% filter(FishingType %in% c("Casting", "Demersal"))
    
    BR <- readRDS("./data/gpkg/2.1_BRtrips.gpkg") %>% st_transform(crs = 4283)# from BRtrips.R
    
    hol <- read_csv("./data/RAW/Holidays.csv") %>% 
      mutate(Date = as.POSIXct(Date, format = "%d/%m/%y"))
    
    boat_perf <- read_csv("./data/RAW/boatengines.csv")
    
    fuel <- read_csv("./data/RAW/Monthly-ULP-prices-Gascoyne-201507-202109.csv")
    
    vott <- read.csv("./data/RAW/VOTT.csv")
    
    # models
    cons_mod <- readRDS("./models/cons_mod.rda")
    sp_mod <- readRDS("./models/sp_mod.rda")
    
    source("./functions/genFunc.R") # function for smple weighting
    source("./functions/spatialFunc.R") # function for smple weighting
    
     sf_use_s2(FALSE)
```

# Joining external data
# Side/launch/season
```{r}
dat %<>%
  mutate(facYear = as.factor(numYear)) %>% 
  mutate(side = ifelse(Site %in% c("ExmouthBR", "BundegiBR"), "East", "West")) %>% 
  mutate(side = as.factor(side)) %>% 
  mutate(launch2 = ifelse(Site %in% c("ExmouthBR", "BundegiBR"), "G", NA)) %>% 
  mutate(launch2 = ifelse(Site %in% "TantabiddiBR", "T", launch2)) %>% 
  mutate(launch2 = ifelse(is.na(launch2), "CB", launch2)) %>% 
  mutate(launch2 = as.factor(launch2)) %>% 
  mutate(season = ifelse(TripMonth %in% c("Sep", "Oct"), "Spring", 
                           ifelse(TripMonth %in% c("Jul", "Aug"), "Winter",
                                  "Autumn"))) 

```


## Fuel
```{r}
fuel %<>%
  mutate(TripMonth = gsub("[^a-zA-Z]", "", Month)) %>%
  mutate(season = ifelse(TripMonth %in% c("Sep", "Oct"), "Spring", 
                           ifelse(TripMonth %in% c("Jul", "Aug"), "Winter",
                                  "Autumn"))) %>% 
  mutate(season = as.factor(season)) %>% 
  mutate(numYear = as.numeric(paste0("20", as.numeric(gsub("[^0-9]", "", Month))))) %>%
  mutate(cost = Average/100)

dat %<>% left_join(fuel[, c("TripMonth", "numYear", "cost")])

rm(fuel)
```

## Boat specs
```{r}
# Boat Length - filling in NAs with model (not worth it)
# dat %<>% rename(PartySize = Party)
# tmp <- dat %>% filter(is.na(BoatLength) & SiteType == "Boat") %>% dplyr::select(TripID, BoatLength, PartySize)
# unique(tmp$Party)

# Filling Party size where it exists
# tmp %<>% mutate(PartySize = ifelse(PartySize %in% c("6f", "Two couples 2 kids", "3F, 2M, 1b"), 6, 
#                                     ifelse(PartySize %in% c("2f", "2 dudes", "2males", "Couple", "1M 1b"), 2, 
#                                            ifelse(PartySize %in% c("2f, 2kids", "4 friends", "2 couples", "4 mates", "1m1b", "Couple, 2g", "Couple,2g 1b"), 4,
#                                                   ifelse(PartySize %in% c("Man", "1M"), 1, 
#                                                          ifelse(PartySize %in% c("2b1g", "3M", "3m"), 3, NA))))))

      # tmp$lm_bl <- predict(bl_mod, tmp) # predicting boatlength using party size
      # tmp$lm_bl <- round(as.numeric(tmp$lm_bl), 2)
      # dat %<>%  left_join(tmp[, c("TripID", "lm_bl")])
      
      # dat %<>% 
      #   mutate(BoatLength = ifelse(SiteType == "Boat" & !is.na(lm_bl), lm_bl, BoatLength)) %>% # appending modelled boat length
      #   mutate(BoatLength = ifelse(SiteType == "Boat" & is.na(BoatLength), mean(BoatLength, na.rm = T), BoatLength)) # filling rest of boat length with NAs
        
 # Consumption
boat_perf %<>%  
  filter(kmperh < 4000) %>% 
    mutate(boat_size2L = ifelse(boatlength <= 3.75, "S", "L")) %>%
  mutate(boat_size3L = ifelse(boatlength <= 3.75, "S", 
                            ifelse(boatlength <= 6.75, "M", "L"))) %>% 
  mutate(cons = kmperl/nEngines)

boat3L <- boat_perf %>% 
  group_by(boat_size3L) %>% 
  summarise(sp3l = median(kmperh),
            cons3l = median(cons))

boat2L <- boat_perf %>% 
  group_by(boat_size2L) %>% 
  summarise(sp2l = median(kmperh),
            cons2l = median(cons))
  
dat %<>% 
   mutate(BoatLength = ifelse(is.na(BoatLength), mean(BoatLength, na.rm = TRUE), BoatLength)) %>%
  mutate(boat_size2L = ifelse(BoatLength <= 3.75, "S", "L")) %>%
  mutate(boat_size3L = ifelse(BoatLength <= 3.75, "S", 
                            ifelse(BoatLength <= 6.75, "M", "L"))) %>% 
  left_join(boat2L[, c("boat_size2L", "cons2l", "sp2l")], by = "boat_size2L") %>% 
  left_join(boat3L[, c("boat_size3L", "cons3l", "sp3l")], by = "boat_size3L")

dat %<>% 
  mutate(cons_flt = median(boat_perf$cons)) %>% 
  mutate(sp_flt = median(boat_perf$kmperh))
  

# modeled consumption
 tmp <- dat %>% 
        dplyr::select(BoatLength) %>% 
        rename(boatlength = BoatLength) %>% 
        st_drop_geometry()
      
      gam_cons <- predict(cons_mod, tmp)
      gam_sp <- predict(sp_mod, tmp)
      
      dat$gam_cons <- as.numeric(gam_cons)
      dat$gam_sp <- as.numeric(gam_sp)
      
rm( boat_perf, boat2L, boat3L, gam_cons, cons_mod, sp_mod, gam_sp)
```

## VOTT
```{r}
# Append census data
    dat %<>% left_join(vott[, c("Postcode", "ann_inc_21", "VOTT_21")]) 

# Model income ~ boat size to fill NAs
## Make small modelling data set
    tmp <- dat %>% 
      filter(!is.na(ann_inc_21) & !is.na(BoatLength)) %>% 
      mutate(ann_inc_21 = as.numeric(ann_inc_21))

## Linear model
    lm_ann_inc <- lm(ann_inc_21 ~ BoatLength, data = tmp)

## Check 
      ggplot(dat = tmp, aes(x = ann_inc_21, y = BoatLength)) +
        geom_point() +
        stat_smooth(method = "lm")

## Predict to data 
      dat$lm_ann_inc <- predict(lm_ann_inc, dat)

# Calculate vott from annual income
      dat %<>%
        mutate(ann_inc = ifelse(!is.na(ann_inc_21), ann_inc_21, lm_ann_inc)) %>% 
        mutate(ann_inc = ifelse(is.na(ann_inc), mean(ann_inc, na.rm = T), ann_inc)) %>% 
        mutate(vott = round((ann_inc/(38*52)/3), 2))


rm(tmp, lm_ann_inc, vott)
```

# Filtering
All data filters have to be done before making the choice set.
```{r}
    dat %<>% 
      filter(ActivityType == "Extractive") %>% # filtering activity
      filter(str_detect(Site, "BR")) %>% 
      filter(SiteType %in% "Boat") %>% # filtering activity
      filter(!is.na(UseLat), !is.na(UseLong)) %>% # filtering empty coords
      st_as_sf(coords = c("UseLong", "UseLat"), crs = 4283) %>%  ## make sf object
      rename(dat_geom = geometry)
      
    # need one trip per row (longest site visited on trip)
    dat %<>% 
      group_by(PersonID, TripID) %>% 
      slice(which.max(decDuration)) %>% 
      ungroup() 
    
    dat <- st_crop(dat, grid) # removing outliers 
    
      #'[#VALIDATE: TRUE, no duplicated trip IDs)]
      identical((which(duplicated(dat$TripID) == TRUE)), integer(0)) == TRUE
```

# Weighting
Following protocol outed in KAP

## Step 5.1: Identify starta to be weighted
  - boat ramp (no others necessary because on intensive field trip)
  
The primary sampling unit (psuid) is a shift am/pm at a specific boat ramp on a specific date where at least one survey was conducted
```{r}
## Prepping data 
tot_smpl <- nrow(dat) # total samples

# Append frequency of each starta to data set
br_freq <- table(dat$Site) %>% as.data.frame() %>% rename(br_freq = Freq)

dat %<>% left_join(br_freq, by = c("Site" = "Var1"))
```

## Step 5.2: Calculate inclusion probability
```{r}
# Step 5.2a: calulate inclusion probability
dat %<>%
  mutate(br_inprob = br_freq/tot_smpl)
```

## Step 5.3: Calculate IPW
```{r}
dat %<>% mutate(ipw = 1/br_inprob)
```

# Choice set
```{r choice set}
  # isolating grid ID to append to dat
    gridID_vis <- grid %>%
      dplyr::select(gridid, grid_geom) %>%
      distinct()

which(duplicated(gridID_vis$gridid)) # should be none

# Writing choice set
# choice <- sites %>% dplyr::select(gridID_alt, geom)
# st_write(choice, "./data/gpkg/2.1_ChoiceSetQ_4283.gpkg", append = FALSE) # need to write with st_write to open in QGISRUMs
# saveRDS(sites, "./data/gpkg/2.1_ChoiceSet_4283.gpkg")

# Appending visited grids to data
dat %<>% 
  st_join(gridID_vis, left = T, join = st_intersects) %>% 
  rename(gridID_vis = gridid)

# Selecting data attributes
dat %<>% 
  # rename(gridID_vis = gridID_alt) %>% 
  dplyr::select(TripID, PersonID, gridID_vis, Site, TripDate, MedianTime, BoatLength, cost, ipw, Postcode, cons_flt, sp_flt, cons2l, sp2l, cons3l, sp3l, gam_cons, gam_sp, vott, ann_inc, KeptUndam, CaughtUndam, Dem, Dem_bin,
                Start_hr, BaitLure, Age, exTimes12m, TripMonth, numYear, decDuration, FishingType) %>% 
  distinct() %>% 
  mutate(RampID = ifelse(Site %in% "BundegiBR", "73", NA),
         RampID = ifelse(Site %in% "CoralBayBR", "72", RampID),
         RampID = ifelse(Site %in% "ExmouthBR", "85", RampID),
         RampID = ifelse(Site %in% "TantabiddiBR", "87", RampID)) %>% 
  mutate(UseLong = st_coordinates(dat_geom)[,1]) %>% 
  mutate(UseLat = st_coordinates(dat_geom)[,2]) %>% 
  as.data.frame()

    #'[#VALIDATE: TRUE, there are no unassigned gridIDs]
    
    identical((which(is.na(dat$gridID_vis) == TRUE)), integer(0)) == TRUE
    
 # Make full choice set
    t1 <- nrow(dat) # copying original data test join

    dat <- inner_join(dat, grid, by = "RampID") 
    
    #'[#VALIDATE: TRUE, number of sites * number of trips == nrow(dat), unless sites have been removed eg. beach launching sites]
    
    t1*nrow(gridID_vis) == nrow(dat)
```


# Choice and travel cost ($0.54/km)
```{r choice and travel cost}

      dat <- dat %>%
        rename(gridID_alt = gridid) %>% 
        # filter(!is.na(Postcode)) %>% 
        mutate(choice = ifelse(gridID_vis == gridID_alt, 1, 0)) %>%
        # mutate(land_fc = round(as.numeric((cost/(100/12))*accom_br), 2)*2) %>% 
        # mutate(water_fc = round(as.numeric((cost/consump)*km_BR), 2)*2) %>% #changing fuel&cost
        # mutate(lm_fc = round(as.numeric((cost/lm_cons)*km_BR), 2)*2) %>% 
        
  # fuel cost
        mutate(l_fc_flt = round(as.numeric((cost/cons_flt)*l_km_br), 2)) %>% 
        mutate(nl_fc_flt = round(as.numeric((cost/cons_flt)*nl_km_br), 2)) %>% 
        mutate(l_fc2l = round(as.numeric((cost/cons2l)*l_km_br), 2)) %>% 
        mutate(nl_fc2l = round(as.numeric((cost/cons2l)*nl_km_br), 2)) %>% 
        mutate(l_fc3l = round(as.numeric((cost/cons3l)*l_km_br), 2)) %>% 
        mutate(nl_fc3l = round(as.numeric((cost/cons3l)*nl_km_br), 2)) %>%
        mutate(l_fc_gam = round(as.numeric((cost/gam_cons)*l_km_br), 2)) %>% 
        mutate(nl_fc_gam = round(as.numeric((cost/gam_cons)*nl_km_br), 2)) %>%
  
# travel time
        mutate(l_trvlTime_flt = round(as.numeric(l_km_br/sp_flt, 2))) %>%
        mutate(nl_trvlTime_flt = round(as.numeric(nl_km_br/sp_flt, 2))) %>%
        mutate(l_trvlTime2l = round(as.numeric(l_km_br/sp2l, 2))) %>%
        mutate(nl_trvlTime2l = round(as.numeric(nl_km_br/sp2l, 2))) %>%
        mutate(l_trvlTime3l = round(as.numeric(l_km_br/sp3l), 2)) %>%
        mutate(nl_trvlTime3l = round(as.numeric(nl_km_br/sp3l), 2)) %>%
        mutate(l_trvlTime_gam = round(as.numeric(l_km_br/gam_sp), 2)) %>%
        mutate(nl_trvlTime_gam = round(as.numeric(nl_km_br/gam_sp), 2)) %>%
  
  # trip VOTT
        mutate(l_trp_vott_flt = round(as.numeric(vott*l_trvlTime_flt), 2)*2) %>%
        mutate(nl_trp_vott_flt = round(as.numeric(vott*nl_trvlTime_flt), 2)*2) %>%
        mutate(l_trp_vott2l = round(as.numeric(vott*l_trvlTime2l), 2)*2) %>%
        mutate(nl_trp_vott2l = round(as.numeric(vott*nl_trvlTime2l), 2)*2) %>%
        mutate(l_trp_vott3l = round(as.numeric(vott*l_trvlTime3l), 2)*2) %>%
        mutate(nl_trp_vott3l = round(as.numeric(vott*nl_trvlTime3l), 2)*2) %>%
        mutate(l_trp_vott_gam = round(as.numeric(vott*l_trvlTime_gam), 2)*2) %>%
        mutate(nl_trp_vott_gam = round(as.numeric(vott*nl_trvlTime_gam), 2)*2) %>%
  
  # total cost
        mutate(l_tc_flt = round(as.numeric(l_fc_flt + l_trp_vott_flt), 2)) %>% 
        mutate(nl_tc_flt = round(as.numeric(nl_fc_flt + nl_trp_vott_flt), 2)) %>%
        mutate(l_tc2l = round(as.numeric(l_fc2l + l_trp_vott2l), 2)) %>% 
        mutate(nl_tc2l = round(as.numeric(nl_fc2l + nl_trp_vott2l), 2)) %>% 
        mutate(l_tc3l = round(as.numeric(l_fc3l + l_trp_vott3l), 2)) %>% 
        mutate(nl_tc3l = round(as.numeric(nl_fc3l + nl_trp_vott3l), 2)) %>% 
        mutate(l_tc_gam = round(as.numeric(l_fc_gam + l_trp_vott_gam), 2)) %>% 
        mutate(nl_tc_gam = round(as.numeric(nl_fc_gam + nl_trp_vott_gam), 2)) %>% 
  
  # differential total cost (always flat fuel cost)
        mutate(l_fcflt_sp2l = round(as.numeric(l_fc_flt + l_trp_vott2l), 2)) %>% 
        mutate(nl_fcflt_sp2l = round(as.numeric(nl_fc_flt + nl_trp_vott2l), 2)) %>% 
        mutate(l_fcflt_sp3l = round(as.numeric(l_fc_flt + l_trp_vott3l), 2)) %>% 
        mutate(nl_fcflt_sp3l = round(as.numeric(nl_fc_flt + nl_trp_vott3l), 2)) %>% 
        mutate(l_fcflt_spgam = round(as.numeric(l_fc_flt + l_trp_vott_gam), 2)) %>% 
        mutate(nl_fcflt_spgam = round(as.numeric(nl_fc_flt + nl_trp_vott_gam), 2)) 
        
 #'[#VALIDATE: TRUE, should equal number of trips]
    sum(dat$choice) == n_distinct(dat$TripID)
```

# Transformations
```{r}
    dat$arealog <- log(dat$area + 1)
    dat$log_nl_km_br <- log(dat$nl_km_br + 1)
    dat$log_l_km_br <- log(dat$l_km_br + 1)
    dat$log_l_k <- log(dat$l_km_br + 1)
    dat$l_fcflt_aninc <- dat$l_fc_flt * dat$ann_inc
    dat$nl_fcflt_aninc <- dat$nl_fc_flt * dat$ann_inc
    dat$bl_l_offdist <- dat$BoatLength * dat$log_l_km_br
    dat$bl_nl_offdist <- dat$BoatLength * dat$log_nl_km_br
```


## Swell
```{r}
    dat <- dat %>% 
      mutate(centLong = st_coordinates(centroid)[,1]) %>% # need this is using dynamic variables
      mutate(centLat = st_coordinates(centroid)[,2])

  #extracting swell
   hsws <- get_hs_ws_day(dat$TripDate, dat$centLong, dat$centLat) # getting hs and ws

   # making a numeric hour column
  dat$MedianHour <- as.numeric(gsub("\\:.*$", "", dat$MedianTime))

  # catagorising time of day
  dat$TimeOfDay <- with(dat,  ifelse(MedianHour >= 5 & MedianHour <= 10, "Morning",
                            ifelse(MedianHour >= 11 & MedianHour <= 13, "Midday",
                                   ifelse(MedianHour >= 14 & MedianHour <= 18, "Afternoon", NA))))

  # appending swell and wind speed
  hsws_df <- as.data.frame(hsws)

  dat <- dat %>%
    mutate(swell = ifelse(TimeOfDay %in% "Morning", hsws$hs_morning,
                                ifelse(TimeOfDay %in% "Midday", hsws$hs_midday, hsws$hs_afternoon))) %>%
    mutate(wind = ifelse(TimeOfDay %in% "Morning", hsws$wspeed_morning,
                                ifelse(TimeOfDay %in% "Midday", hsws$wspeed_midday,
                                       hsws$wspeed_afternoon)))
```
# Save output
```{r save}
dat_rum_df <- dat %>% dplyr::select(-c(centroid, br_geom, grid_geom, dat_geom))

dat_rum_df %<>% mutate(rug = as.numeric(rug)) # add this to static script

  write_csv(dat_rum_df, "./data/02_data/2.1_dat.csv")
  saveRDS(dat, paste0("./data/gpkg/sims/2.1_dat.gpkg")) # script 3.1

# saveRDS(dat, "./data/gpkg/2.1_dat.gpkg")
```
