==> 2.2_DynamicCov.Rmd <==
```{r setup}
rm(list = ls())

# knitr options
knitr::opts_chunk$set(warning = FALSE, message = FALSE, echo = FALSE, fig.align = 'center', fig.width = 10, fig.height = 7) 

# libraries 
library(tidyverse)
library(dplyr)
library(ggplot2)
library(sp)
library(raster)
library(rgeos)
library(rgdal)
library(sf)
library(RNetCDF)
library(ncdf4)
library(googledrive)
library(corrplot)
# library(windfetch)

# functions
source("./functions/spatialFunc.R") # spatial functions

# data
dat <- read_csv("./data/02_data/2.1_dat.csv")

## spatial
### gpkg - ESPG 4283 (GDA94)
# sites <- readRDS("./data/gpkg/2.0_Sites_4283.gpkg") # sites with static attributes
coast <- st_read("./data/gpkg/coast.gpkg") %>%  st_transform(4283)
```

# Data prep
```{r data prep}
# # making use data spatial 
# dat %<>%
#   filter(!is.na(UseLat), !is.na(UseLong)) %>%
#   st_as_sf(coords = c("UseLong", "UseLat"), crs = 4283)

# dat %<>%
#   filter(!is.na(centroidLat), !is.na(centroidLong)) %>%
#   st_as_sf(coords = c("centroidLong", "centroidLat"), crs = 4283)

# sm_dat <- dat %>%  
#   mutate(trp_grid = str)
```

# Enviromental variables
## SST
```{r}
# Make small data to extract sst - test if this makes it quicker
# tmp <- dat %>% 
#   dplyr::select(gridid_alt, TripDate, centroidLong, centroidLat) %>%
#   distinct_at(vars("gridid_alt", "TripDate"), .keep_all = TRUE)

coast_lambert <- coast %>%
  st_transform(crs = 3112) # need land lambert shp file loaded

sst_kelvins <- get_sst(dat$TripDate, Long = dat$centroidLong, Lat = dat$centroidLat) # getting sst

sst_celcius <- round(sst_kelvins - 273.15, 2) # convert to celcius

dat$sst <- sst_celcius

# tmp <- dat %>%  filter(is.na(sst))
# 
# t <- tmp %>% distinct(gridid_alt, .keep_all = TRUE)
# 
# t %<>%
#   filter(!is.na(centroidLat), !is.na(centroidLong)) %>%
#   st_as_sf(coords = c("centroidLong", "centroidLat"), crs = 4283)
# 
# ggplot() +
#   geom_sf(data = coast) +
#   geom_sf(data = t, colour = 'red')

# sst missing for one person 2020-09-25, trip 2
# filling with data from day after

tmp2 <- dat %>% 
  filter(TripDate == "2020-09-26") %>% 
  distinct(gridid_alt, .keep_all = T) %>% 
  dplyr::select(gridid_alt, sst) %>% 
  rename(sst2 = sst)

dat %<>% left_join(tmp2)

dat %<>% mutate(sst = is.na(sst), sst2, sst) 
# which(is.na(dat$sst))

# Missing data is centroid that are on land - dont think ill need it
```

## Swell height and wind speed
Data is extracted for morning: 8-10, midday: 11-1 and afternoon: 2-4. 
Not very high res - not useful for area whith lots of islands like Esperance. 
```{r hs and ws}
hsws <- get_hs_ws_day(dat$TripDate, dat$centroidLong, dat$centroidLat) # getting hs and ws

# making a numeric hour column
dat$MedianHour <- as.numeric(gsub("\\:.*$", "", dat$MedianTime))

# catagorising time of day
dat$TimeOfDay <- with(dat,  ifelse(MedianHour >= 5 & MedianHour <= 10, "Morning",
                            ifelse(MedianHour >= 11 & MedianHour <= 13, "Midday",
                                   ifelse(MedianHour >= 14 & MedianHour <= 18, "Afternoon", NA))))
# spdat[, c("MedianTime", "MedianHour", "TimeOfDay")] # check it has worked

hsws_df <- as.data.frame(hsws)

dat %<>%
  mutate(swell = ifelse(TimeOfDay %in% "Morning", hsws$hs_morning,
                              ifelse(TimeOfDay %in% "Midday", hsws$hs_midday, hsws$hs_afternoon))) %>%
  mutate(wind = ifelse(TimeOfDay %in% "Morning", hsws$wspeed_morning,
                              ifelse(TimeOfDay %in% "Midday", hsws$wspeed_midday,
                                     hsws$wspeed_afternoon)))
```

# Save output
```{r save}
# save a dynamic vars frame to be joined on later if you need
dyn_cov <- dat %>% dplyr::select(gridid_alt, TripID, sst, swell, wind)
# write_csv(dyn_cov, "./data/02_data/2.2_dyn_cov_base.csv")
# write_csv(dyn_cov, "./data/02_data/2.2_dyn_cov_sim.csv")

# save with geoms
# saveRDS(dat, "./data/02_data/2.2_dat_base.gpkg")
# saveRDS(dat, "./data/02_data/2.2_dat_sim.gpkg")

# data frame with no geom for stata
dat_df <- dat %>%  st_drop_geometry %>% as.data.frame()
# write.csv(dat, "./data/02_data/2.2_dat_base.csv")
# write.csv(dat, "./data/02_data/2.2_dat_sim.csv")
```
