==> 2.3_ModelCov.Rmd <==
# Setup
```{r setup}
rm(list = ls())

# knitr options
knitr::opts_chunk$set(warning = FALSE, message = FALSE, echo = FALSE, fig.align = 'center', fig.width = 10, fig.height = 7) 

# libraries
# library(ggBRT)
library(gbm)
library(dismo)
library(MASS)
library(dplyr)
library(gam)
library(mgcv)
library(ggplot2)
library(FSSgam)
library(corrplot)
library(corrr)
library(tidyverse)
library(magrittr)
library(sf)
library(car)
library(doBy)
require(gplots)
library(RColorBrewer)
library(janitor)
# library(ggBRT)

# functions
source("./functions/brtFunc.R")
source("./functions/theme.R")

# Simulation
sim_name <- "sim3"

#### Full data - ctcah will be predicted to this data 
full_dat <- readRDS("./data/gpkg/sims/2.1_dat.gpkg") %>%  # ATTRIBUTE MODEL
  rename(catch = CaughtUndam) %>% # making stat friendly
  mutate(side = ifelse(Site %in% c("ExmouthBR", "BundegiBR"), "East", "West")) %>%
  mutate(side = as.factor(side)) %>%
  mutate(launch2 = ifelse(Site %in% c("ExmouthBR", "BundegiBR"), "G", NA)) %>%
  mutate(launch2 = ifelse(Site %in% "TantabiddiBR", "T", launch2)) %>%
  mutate(launch2 = ifelse(is.na(launch2), "CB", launch2)) %>%
  mutate(launch2 = as.factor(launch2)) %>% 
  mutate(season = ifelse(TripMonth %in% c("Sep", "Oct"), "Spring", 
                           ifelse(TripMonth %in% c("Jul", "Aug"), "Winter",
                                  "Autumn"))) %>%
  mutate(facyear = as.factor(numYear))

# full_dat <- read.csv("./data/02_data/2.1_dat_base.csv") # CHARLOTTE - NO SIM

# full_dat <- readRDS(paste0("./data/gpkg/sims/2.1_ascCS_rum_", sim_name,"_4283.gpkg")) %>%  # ASC MODEL
#   # Ad this to wrangle next time you run it. 
#   mutate(facYear = as.factor(numYear)) %>% 
#   mutate(side = ifelse(Site %in% c("ExmouthBR", "BundegiBR"), "East", "West")) %>% 
#   mutate(side = as.factor(side)) %>% 
#   mutate(launch2 = ifelse(Site %in% c("ExmouthBR", "BundegiBR"), "G", NA)) %>% 
#   mutate(launch2 = ifelse(Site %in% "TantabiddiBR", "T", launch2)) %>% 
#   mutate(launch2 = ifelse(is.na(launch2), "CB", launch2)) %>% 
#   mutate(launch2 = as.factor(launch2)) %>% 
#   mutate(logdepth = log(depth)) %>% 
#   rename(catch = CaughtUndam) %>% # making stat friendly
#   mutate(season = ifelse(TripMonth %in% c("Sep", "Oct"), "Spring", 
#                            ifelse(TripMonth %in% c("Jul", "Aug"), "Winter",
#                                   "Autumn"))) 

# making cols stat friendly
names(full_dat) <- tolower(names(full_dat))
names(full_dat) <- gsub("\\.", "", names(full_dat))
          
##### Filter full data to data to use for modelling
dat <- full_dat %>% filter(choice == 1 & fishingtype %in% c("Demersal"))
                           # # & Site %in% c("TantabiddiBR", "CoralBayBR"))
                           # & Site %in% c("BundegiBR", "ExmouthBR") 
                           # & facYear %in% c("2015", "2016"))
```
Stata will change all vars to lower case and will remove full stops. Please make all vars stata friendly and do transformations first. 


# Responses
Data avaliable
Jon's data: catch (numeric) and fishing type (demersal or trolling)
Nicole's data: catch, fishing type and fish group caught (e.g. demersal, pelagic etc.)

I am going to store the optimal catch predictions for the following responses options, ordered by what seems most sensible ot suse:

  1. catch (previously CaughtUndam): all data, continuous response, consistent between Nicole's and Jon's data (based on fishing type only)
  2. catch_bin: all data, binomial response, consistent between Nicole's and Jon's data (based on fishing the only)
  3. dem: all data, continuous response, Jon's catch data for demersal fishers and Nicole's records of demersal catch by demersal fishers
  4. dem bin: binomial response, Jon's catch data for demersal fishers and Nicole's records of demersal catch by demersal fishers
  5. average catch

I will test all these models in my ASC + ATT RUM. 

```{r}
# prep responses
dat %<>% 
  mutate(catch_bin = as.factor(ifelse(catch > 0, "1", "0"))) %>% # binary response for caught undam
  mutate(dem = ifelse(numyear < 2019, catch, dem)) %>% # Get Jons data into the demersal column to make dem_bin
  mutate(dem_bin = as.factor(ifelse(dem > 0, "1", "0")))  #  binary response for demersal fish (caught Jon, demersal mine) 

av <- dat %>%
  group_by(gridid_alt) %>%
  summarize(av_catch = round(mean(catch)))

dat %<>% left_join(av)

resp.vars <- c("catch", "dem", "catch_bin", "dem_bin", "av_catch") # define responses

colSums(is.na(dat[, resp.vars])) # make sure no NAs

# Create an empty list to store plots
plots_list <- list()

# Loop through each column in resp.vars
for (col_name in resp.vars) {
  # Check if the column is a factor
  if (is.factor(dat[[col_name]])) {
    # Create a bar plot for factors
    bar_plot <- ggplot(data = dat, aes_string(x = col_name)) +
      geom_bar(fill = "darkturquoise") +
      xlab(col_name) +
      ylab("Frequency") +
      theme_classic()
    
    # Add the bar plot to the list
    plots_list[[col_name]] <- bar_plot
  } else {
    # Calculate percentage of zeros for numeric variables
    zeros <- round(sum(dat[[col_name]] == 0, na.rm = TRUE) / nrow(dat) * 100, 2)
    
    # Create histogram plot for numeric variables
    hist_plot <- ggplot(data = dat, aes_string(x = col_name)) +
      geom_histogram(fill = "darkturquoise") +
      xlab(col_name) +
      ylab("Frequency") +
      annotate("text", x = 0, y = 10, label = paste0(zeros, "% zeros"), angle = 90) +
      theme_classic()
    
    # Add the histogram plot to the list
    plots_list[[col_name]] <- hist_plot
  }
}

print(plots_list)

# 1. catch: 3% zeros, but lots of low numbers, if you look at catch_bin you can see that most people catch at least one fish - this is because there is nearshore fish mixed in wit this.  
# 2. catch_bin: more people catch fish than don't catch because of near shore fish mixed in - not ideal (maybe dem_bin better)
# 3. dem: same pattern as catch but higher % zeros (8%)
# 4. dem_bin: as expected, most people dont catch fish
# 5. av_catch: distribution looks terrible, but might be more spatially accuarte

# testing spatial distribution of average catch
grid <- dat %>% distinct(gridid_alt, .keep_all = TRUE)

ggplot() +
  geom_sf(dat = grid$asc_geom, aes(fill = grid$av_catch))

# some high catch nearshore, low catch in AMP near boat ramps
```


# FSSgam
## Covariates to test
```{r}
# set the continuous predictors to test
# pred.vars <- c("dem", "depth", "l_km_br", "nl_km_br", "pelagic", "reef", "rreef", "lagoon", "decDuration", "rug", "centroidLat") # FOR CHARLOTTE
# pred.vars <- c("depth", "km_br", "reefs", "lagoon", "rug", "medianhour", "extimes12m", "centlat") # ASC
pred.vars <- c("depth", "reefs", "lagoon", "rug", "centlat", "medianhour", "extimes12m") # ATTRIBUTE

# factor vars 
factor.vars <- c("season", "baitlure", "launch2", "side") 
dat %<>% mutate(across(all_of(c(factor.vars)), as.factor))

# rum.vars
# rum.vars <- c("fcflt_spgam")
rum.vars <- c("nl_fcflt_spgam")
```

## NAs
There needs to be no NAs in the predictors in the full data set the model can predict on to the full data set. 
```{r}
# check continuous vars 
summary(full_dat[,pred.vars]) 
pred.vars <- pred.vars[ !pred.vars == "extimes12m"] # remove 

# check factors vars
names(which(colSums(is.na(dat[,factor.vars]))>0)) 
factor.vars <- factor.vars[ !factor.vars == "baitlure"] # remove
```

## Correlation
Check the correlation between the catch and rum predictors. 

Concerned about correlation between latitude, side and launch. Latitude was always included but we added side because latitude in the are and west are different. Launch is basically the boat ramp so included both the side and launch so i think i need to have side & latitude OR launch. However, now that i think about it, i think launch might be correlated with cost?

Removing launch as a variable because of potential correlation with cost. 

```{r}
# Check the correlations between predictor variables - remove if over 95
correlate(dat[,c(pred.vars, rum.vars)], use = "complete.obs") %>%  
  gather(-term, key = "colname", value = "cor") %>% 
  dplyr::filter(abs(cor) > 0.5) %>% 
  arrange(-cor) %>% 
  distinct(cor, .keep_all = TRUE)

pred.vars <- pred.vars[ !pred.vars == "km_br"] # remove high correlations

# I think launch and side/lat are correlated and need to be eitehr or 
# factor.vars <- factor.vars[ !factor.vars == "launch2"] # test without launch
factor.vars <- factor.vars[ !factor.vars == "side"] # test without launch
# pred.vars <- pred.vars[ !pred.vars == "centlat"] # test without launch
```


## Outliers
Check for outrageous outliers that need to be removed from the data. 
```{r}
par(mfrow = c(1,2))
for (i in pred.vars) {
  x <- dat[,i]
  x = as.numeric(unlist(x))
  hist((x), main = paste(i))
  plot((x), main = paste(i))
}

dat %<>% filter(depth < 600) # removing one observation at 900m
```

## Tranformations
Aiming for homoscedascit, normal distribution, however avoid transformation where possible. 
```{r}
  par(mfrow = c(3,4))

  for (i in pred.vars) {
    
    tryCatch({
      
       x <- dat[,i]
       x = as.numeric(unlist(x))
       hist((x))
       plot((x), main = paste(i))
       hist(sqrt(x))
       plot(sqrt(x))
       hist(log(x + 1))
       plot(log(x + 1))
       hist((x))
       plot((x), main = paste(i))
       hist((x)^2)
       plot((x)^2)
       hist((x)^3)
       plot((x)^3)
       
    }, error = function(e) {
      # Handle the error if necessary
      print(paste("Unable to transform", i, ":", e))
    }
    )
  }
  
  # dat$log_lkm_br <- log(dat$l_km_br + 1) # CHARLOTTE
  # dat$log_nlkm_br <- log(dat$nl_km_br + 1) # CHARLOTTE
```

## Assign pred.vars
```{r}
# pred.vars <- c("log_depth","log_nlkm_br","pelagic", "log_reef", "log_rreef", "log_lagoon", "sqrt_decDuration", "sqrt_rug", "centroidLat") # CHARLOTTE
# pred.vars <- c("log_depth","log_nl_km_br", "log_reef", "log_rreef", "log_lagoon", "sqrt_rug", "centLat", "log_avid", "MedianHour") # Attribute
```

## Model selection
  - k is knows determines out wavy we allow the line to be - the higher the knots the smoother and more risk of over fitting: max 5
  - cr is a cubic spline
  - re = random effect

```{r}
# savedir <- "./models/catch/without_launch" 
# savedir <- "./models/catch/with_launch" 
savedir <- "./models/att_catch/with_launch" # attribute
use.dat <- as.data.frame(dat) 
out.all <- list()
var.imp <- list()

for (i in 1:length(resp.vars)) {
  print(resp.vars[i])
  
  if (is.factor(use.dat[[resp.vars[i]]])) {
    # If response variable is binary (factor), use binomial distribution
    dist_family <- binomial()
  } else {
    # If response variable is continuous, use tweedie distribution
    dist_family <- tw()
  }
  
  Model1 <- gam(as.formula(paste(resp.vars[i], "~ s(depth, k = 3, bs='cr') + s(facyear, bs = 're')")), 
                 offset = decduration, family = dist_family, data = use.dat)
  
  model.set <- generate.model.set(use.dat = use.dat,
                                  test.fit = Model1,
                                  pred.vars.cont = pred.vars,
                                  pred.vars.fact = factor.vars,
                                  max.predictors = 5,
                                  k = 3,
                                  null.terms = "s(facyear, bs = 're')",
                                  smooth.smooth.interactions = FALSE,
                                  factor.factor.interactions = FALSE,
                                  cov.cutoff =  TRUE)
  
  out.list <- fit.model.set(model.set,
                            max.models = 600,
                            parallel = TRUE)
  
  names(out.list)
  out.list$failed.models # examine the list of failed models
  
  mod.table <- out.list$mod.data.out  # look at the model selection table
  mod.table <- mod.table[order(mod.table$AICc), ]
  mod.table$cumsum.wi <- cumsum(mod.table$wi.AICc)
  out.i   <- mod.table[which(mod.table$delta.AICc <= 2), ]
  out.all <- c(out.all, list(out.i))
  var.imp <- c(var.imp, list(out.list$variable.importance$aic$variable.weights.raw)) #Or importance score weighted by r2
  
  # plot the best models
  # for (m in 1:nrow(out.i)) {
  #   best.model.name <- as.character(out.i$modname[m])
  #   png(file = paste(savedir, paste(sim_name, "_", resp.vars[i], "mod_fits.png", sep = "_"), sep = "/"))
  #   if (best.model.name != "null") {
  #     par(mfrow = c(3, 1), mar = c(9, 4, 3, 1))
  #     best.model <- out.list$success.models[[best.model.name]]  # something wrong with hoe best.model is defined - cant find in list
  #     plot(best.model, all.terms = TRUE, pages = 1, residuals = TRUE, pch = 16)
  #     mtext(side = 2, text = resp.vars[i], outer = FALSE)
  #   }
  #   dev.off()
  # }
}

names(out.all) <- resp.vars
names(var.imp) <- resp.vars
all.mod.fits   <- do.call("rbind",out.all)
all.var.imp    <- do.call("rbind",var.imp)
write.csv(all.mod.fits[ , -2], file = paste(savedir, paste(sim_name, "all.mod.fits.csv", sep = "_"), sep = "/"))
write.csv(all.var.imp, file = paste(savedir, paste(sim_name, "all.var.imp.csv", sep = "_"), sep = "/"))

# importance plot
pdf(file = paste0("./models/catch/", sim_name, "_var_importance.pdf"), height=5, width=7, pointsize=10)
heatmap(all.var.imp, notecex = 0.4, col = colorRampPalette(c("yellow","orange","red"))(30), trace = "none", key.title = "", keysize = 2, notecol = "black", key = T,
        sepcolor = "black", margins = c(12,14), lhei = c(3,10), lwid = c(3,10), Rowv = NA, Colv = NA, scale = 'column')
dev.off()
```

## Top models
Check the top models for each resp and see if they make sense. 

### catch
```{r}
### ASC 
# without luanch
# folder <- "catch"
# launch <- "without_launch"
# catch_mod <- gam(catch ~
#                   s(medianhour, by = season, k = 3, bs = "cr") +
#                   s(reefs, by = side, k = 3, bs = "cr") +
#                   s(rug, by = side, k = 3, bs = "cr") +
#                    season +
#                    side +
#                    s(facyear, bs = 're'),
#                  offset = decduration,
#                  method = "REML",
#                  family = tw(),
#                  data = dat
#                 )

# with launch
# launch <- "with_launch"
# catch_mod <- gam(catch ~
#                   s(medianhour, by = season, k = 3, bs = "cr") +
#                   s(reefs, by = launch2, k = 3, bs = "cr") +
#                   s(depth, by = launch2, k = 3, bs = "cr") +
#                    season +
#                    launch +
#                    s(facyear, bs = 're'),
#                  offset = decduration,
#                  method = "REML",
#                  family = tw(),
#                  data = dat
#                 )

#### ATTRIBUTE
# folder <- "att_catch"
# launch <- "with_launch"
# catch_mod <- gam(catch ~
#                   s(medianhour, by = season, k = 3, bs = "cr") +
#                   s(reefs, by = launch2, k = 3, bs = "cr") +
#                   s(depth, by = launch2, k = 3, bs = "cr") +
#                    season +
#                    launch2 +
#                    s(facyear, bs = 're'),
#                  offset = decduration,
#                  method = "REML",
#                  family = tw(),
#                  data = dat
#                 )


summary(catch_mod)
AIC(catch_mod)

par(mfrow = c(2,2)) 
gam.check(catch_mod)

# Save each plot one by one
for (i in 1:length(catch_mod$smooth)) {
    plot(catch_mod, all.terms = TRUE, residuals = TRUE, pch = 16, shade = TRUE, select = i)
    dev.copy(png, filename = paste0("./models/",folder, "/", launch, "/resp_catch/", sim_name, "_catch_mod_plot_", i, ".png"))
    dev.off()
}

# Predict model onto big data
full_dat <- cbind(full_dat, "pred_catch" = predict.gam(catch_mod, newdata = full_dat, se.fit = T, type = "response"))

# Check summary
summary(full_dat$pred_catch.fit)

ggplot() +
  geom_histogram(data = full_dat, aes(x = pred_catch.fit), fill = "darkturquoise", bins = 8) +
  xlab("Predicted Catch") +
  ylab("Frequency") +
  # annotate("text", x = 0, y = 10, label = paste0(zeros, "% zeros"), angle = 90) +
  theme_classic()

grid <- full_dat %>% filter(tripid == 741) %>% st_as_sf() # trip with highest catch rate
  
ggplot() +
  geom_sf(data = grid$asc_geom, aes(fill = grid$pred_catch.fit), lwd = 0.05) +
  theme_classic()
```

### dem
```{r}
### ASC 
# without luanch
# folder <- "catch"
# launch <- "without_launch"
# dem_mod <- gam(dem ~
#                   s(medianhour, by = season, k = 3, bs = "cr") +
#                   s(reefs, by = side, k = 3, bs = "cr") +
#                   s(rug, by = side, k = 3, bs = "cr") +
#                    season +
#                    side +
#                    s(facyear, bs = 're'),
#                  offset = decduration,
#                  method = "REML",
#                  family = tw(),
#                  data = dat
#                 )

# with launch
# launch <- "with_launch"
# dem_mod <- gam(dem ~
#                   s(depth, by = launch2, k = 3, bs = "cr") +
#                   s(lagoon, by = season, k = 3, bs = "cr") +
#                   s(medianhour, by = season, k = 3, bs = "cr") +
#                    season +
#                    launch2 +
#                    s(facyear, bs = 're'),
#                  offset = decduration,
#                  method = "REML",
#                  family = tw(),
#                  data = dat
#                 )

#### ATTRIBUTE
folder <- "att_catch"
launch <- "with_launch"
dem_mod <- gam(dem ~
                  s(depth, k = 3, bs = "cr") +
                  s(reefs, by = launch2, k = 3, bs = "cr") +
                  s(medianhour, by = season, k = 3, bs = "cr") +
                   season +
                   launch2 +
                   s(facyear, bs = 're'),
                 offset = decduration,
                 method = "REML",
                 family = tw(),
                 data = dat
                )

summary(dem_mod)
AIC(dem_mod)

par(mfrow = c(2,2)) 
gam.check(dem_mod)

# Save each plot one by one
for (i in 1:length(dem_mod$smooth)) {
    plot(dem_mod, all.terms = TRUE, residuals = TRUE, pch = 16, shade = TRUE, select = i)
    dev.copy(png, filename = paste0("./models/",folder, "/", launch, "/resp_dem/", sim_name, "_dem_mod_plot_", i, ".png"))
    dev.off()
}

# Predict model onto big data
full_dat <- cbind(full_dat, "pred_dem" = predict.gam(dem_mod, newdata = full_dat, se.fit = T, type = "response"))

# Check summary
summary(full_dat$pred_dem.fit)

ggplot() +
  geom_histogram(data = full_dat, aes(x = pred_dem.fit), fill = "darkturquoise", bins = 8) +
  xlab("Predicted Catch") +
  ylab("Frequency") +
  # annotate("text", x = 0, y = 10, label = paste0(zeros, "% zeros"), angle = 90) +
  theme_classic()

grid <- full_dat %>% filter(tripid == 741) %>% st_as_sf() # trip with highest catch rate
  
#ASC
# ggplot() +
#   geom_sf(data = grid$asc_geom, aes(fill = grid$pred_dem.fit), lwd = 0.05) +
#   theme_classic()

ggplot() +
  geom_sf(data = grid$grid_geom, aes(fill = grid$pred_dem.fit), lwd = 0.05) +
  theme_classic()
```

### catch_bin
```{r}
### ASC 
# without luanch
# folder <- "catch"
# launch <- "without_launch"
# catch_bin_mod <- gam(catch_bin ~
#                   s(medianhour, by = side, k = 3, bs = "cr") +
#                   s(lagoon, by = side, k = 3, bs = "cr") +
#                   s(rug, by = season, k = 3, bs = "cr") +
#                    season +
#                    side +
#                    s(facyear, bs = 're'),
#                  offset = decduration,
#                  method = "REML",
#                  family = binomial(),
#                  data = dat
#                 )

# with launch
# launch <- "with_launch"
# catch_bin_mod <- gam(catch_bin ~
#                   s(medianhour, by = launch2, k = 3, bs = "cr") +
#                   s(rug, by = launch2, k = 3, bs = "cr") +
#                   s(reefs, by = season, k = 3, bs = "cr") +
#                    launch2 +
#                    season +
#                    s(facyear, bs = 're'),
#                  offset = decduration,
#                  method = "REML",
#                  family = binomial(),
#                  data = dat
#                 )

## ATTRIBUTE
folder <- "att_catch"
launch <- "with_launch"
catch_bin_mod <- gam(catch_bin ~
                  s(depth, by = launch2, k = 3, bs = "cr") +
                  s(rug, by = launch2, k = 3, bs = "cr") +
                  s(reefs, by = season, k = 3, bs = "cr") +
                   launch2 +
                   season +
                   s(facyear, bs = 're'),
                 offset = decduration,
                 method = "REML",
                 family = binomial(),
                 data = dat
                )

summary(catch_bin_mod)
AIC(catch_bin_mod)

par(mfrow = c(2,2)) 
gam.check(catch_bin_mod)

# Store effects plots - check to make sure they make sense
png(filename = paste0("./models/", folder, "/",launch, "/resp_catch_bin/", sim_name, "_catch_bin_mod_plot.png"), width = 800, height = 600) # Adjust width and height as needed
plot(catch_bin_mod, trans = plogis, pages = 1, seWithMean = T)
dev.off() # Close the PNG device

# Predict model onto big data
full_dat <- cbind(full_dat, "pred_catch_bin" = predict.gam(catch_bin_mod, newdata = full_dat, se.fit = T, type = "response"))

ggplot() +
  geom_histogram(data = full_dat, aes(x = pred_catch_bin.fit), fill = "darkturquoise", bins = 8) +
  xlab("Predicted Catch") +
  ylab("Frequency") +
  # annotate("text", x = 0, y = 10, label = paste0(zeros, "% zeros"), angle = 90) +
  theme_classic()

grid <- full_dat %>% filter(tripid == 741) %>% st_as_sf() # trip with highest catch rate
  
# ASC
# ggplot() +
#   geom_sf(data = grid$asc_geom, aes(fill = grid$pred_catch_bin.fit), lwd = 0.05) +
#   theme_classic()

#ATTRIBUTE
ggplot() +
  geom_sf(data = grid$grid_geom, aes(fill = grid$pred_catch_bin.fit), lwd = 0.05) +
  theme_classic()
```

### dem_bin
```{r}
### ASC 
# without luanch
# folder <- "catch"
# launch <- "without_launch"
# dem_bin_mod <- gam(dem_bin ~
#                   s(medianhour, by = season, k = 3, bs = "cr") +
#                   s(centlat, k = 3, bs = "cr") +
#                   s(reefs, by = season, k = 3, bs = "cr") +
#                    season +
#                    s(facyear, bs = 're'),
#                  offset = decduration,
#                  method = "REML",
#                  family = binomial(),
#                  data = dat
#                 )

# with launch
# launch <- "with_launch"
# dem_bin_mod <- gam(dem_bin ~
#                   s(depth, by = season, k = 3, bs = "cr") +
#                   s(reefs, by = launch2, k = 3, bs = "cr") +
#                   s(medianhour, by = season, k = 3, bs = "cr") +
#                    season +
#                    launch2 +
#                    s(facyear, bs = 're'),
#                  offset = decduration,
#                  method = "REML",
#                  family = binomial(),
#                  data = dat
#                 )

### ATTRIBUTE
# with luanch
# folder <- "att_catch"
# launch <- "with_launch"
# dem_bin_mod <- gam(dem_bin ~ ## I HAVE NOT RUN THIS MODEL OPTIMAL HAD CENT LAT ALONG WITH LAUNCH WHICH WAS ISTAKE AND IS SHIT ANYWAY
#                   s(depth, by = season, k = 3, bs = "cr") +
#                   s(reefs, by = launch2, k = 3, bs = "cr") +
#                   s(medianhour, by = season, k = 3, bs = "cr") +
#                    season +
#                    launch2 +
#                    s(facyear, bs = 're'),
#                  offset = decduration,
#                  method = "REML",
#                  family = binomial(),
#                  data = dat
#                 )


summary(dem_bin_mod)
AIC(dem_bin_mod)

par(mfrow = c(2,2)) 
gam.check(dem_bin_mod)

# Store effects plots - check to make sure they make sense
png(filename = paste0("./models/catch/", launch, "/resp_dem_bin/", sim_name, "_dem_bin_mod_plot.png"), width = 800, height = 600) # Adjust width and height as needed
plot(dem_bin_mod, trans = plogis, pages = 1, seWithMean = T)
dev.off() # Close the PNG device

# Predict model onto big data
full_dat <- cbind(full_dat, "pred_dem_bin" = predict.gam(dem_bin_mod, newdata = full_dat, se.fit = T, type = "response"))

ggplot() +
  geom_histogram(data = full_dat, aes(x = pred_dem_bin.fit), fill = "darkturquoise", bins = 8) +
  xlab("Predicted Catch") +
  ylab("Frequency") +
  # annotate("text", x = 0, y = 10, label = paste0(zeros, "% zeros"), angle = 90) +
  theme_classic()

grid <- full_dat %>% filter(tripid == 741) %>% st_as_sf() # trip with highest catch rate
  
ggplot() +
  geom_sf(data = grid$asc_geom, aes(fill = grid$pred_dem_bin.fit), lwd = 0.05) +
  theme_classic()
```

### av_catch
```{r}
### ASC 
# without luanch
# folder <- "catch"
# launch <- "without_launch"
# av_catch_mod <- gam(av_catch ~
#                   s(medianhour, k = 3, bs = "cr") +
#                   s(reefs, by = side, k = 3, bs = "cr") +
#                   s(rug, by = side, k = 3, bs = "cr") +
#                    season +
#                    side +
#                    s(facyear, bs = 're'),
#                  offset = decduration,
#                  method = "REML",
#                  family = tw(),
#                  data = dat
#                 )

# with launch
# launch <- "with_launch"
# av_catch_mod <- gam(av_catch ~
#                   s(medianhour, k = 3, bs = "cr") +
#                   s(reefs, by = season, k = 3, bs = "cr") +
#                   s(rug, by = launch2, k = 3, bs = "cr") +
#                    season +
#                    launch2 +
#                    s(facyear, bs = 're'),
#                  offset = decduration,
#                  method = "REML",
#                  family = tw(),
#                  data = dat
#                 )

### ATTRIBUTE
# with luanch
folder <- "att_catch"
launch <- "with_launch"
av_catch_mod <- gam(av_catch ~
                  s(medianhour, k = 3, bs = "cr") +
                  s(lagoon, k = 3, bs = "cr") +
                  s(reefs, by = launch2, k = 3, bs = "cr") +
                  s(depth, by = launch2, k = 3, bs = "cr") +
                   launch2 +
                   s(facyear, bs = 're'),
                 offset = decduration,
                 method = "REML",
                 family = tw(),
                 data = dat
                )

summary(av_catch_mod)
AIC(av_catch_mod)

par(mfrow = c(2,2)) 
gam.check(av_catch_mod)

# Save each plot one by one
for (i in 1:length(av_catch_mod$smooth)) {
    plot(av_catch_mod, all.terms = TRUE, residuals = TRUE, pch = 16, shade = TRUE, select = i)
    dev.copy(png, filename = paste0("./models/", folder, "/", launch, "/resp_av_catch/", sim_name, "_av_catch_mod_plot_", i, ".png"))
    dev.off()
}

# Predict model onto big data
full_dat <- cbind(full_dat, "pred_av_catch" = predict.gam(av_catch_mod, newdata = full_dat, se.fit = T, type = "response"))

# Check summary
summary(full_dat$pred_av_catch.fit)

ggplot() +
  geom_histogram(data = full_dat, aes(x = pred_av_catch.fit), fill = "darkturquoise", bins = 8) +
  xlab("Predicted Catch") +
  ylab("Frequency") +
  # annotate("text", x = 0, y = 10, label = paste0(zeros, "% zeros"), angle = 90) +
  theme_classic()

grid <- full_dat %>% filter(tripid == 741) %>% st_as_sf() # trip with highest catch rate
  
#ASC 
# ggplot() +
#   geom_sf(data = grid$asc_geom, aes(fill = grid$pred_av_catch.fit), lwd = 0.05) +
#   theme_classic()

#Attribute
ggplot() +
  geom_sf(data = grid$grid_geom, aes(fill = grid$pred_av_catch.fit), lwd = 0.05) +
  theme_classic()
```

## Stata friendly
```{r}
# can do this at end to all
full_dat %<>% 
  clean_names() %>%  # replacing . with underscore so it is stata friendly
  mutate_at(vars(starts_with("pred")), round)
```

# Store
Need a data set with no geometry for stata
## ASC
```{r}
# dat_rum_df <- full_dat %>% dplyr::select(-c(asc_geom, br_geom, centroid, dat_geom))
# write_csv(dat_rum_df, paste0("./data/02_data/2.3_asc_dat_", sim_name, ".csv")) # ASC

# ATTRIBUTE
grid %<>% dplyr::select(gridid_alt, br_geom, dat_geom, grid_geom) # spatial data to join later
saveRDS(grid, "./data/gpkg/att_sp_dat.gpkg")
# 
full_dat %<>% dplyr::select(-c(br_geom, dat_geom, grid_geom))
write_csv(full_dat, "./data/02_data/2.3_dat_att.csv")

# CHARLOTTE
# grid %<>% dplyr::select(gridid_alt, br_geom, dat_geom, grid_geom) # spatial data to join later
# saveRDS(grid, "./data/gpkg/att_sp_dat.gpkg")
# 
# full_dat %<>% dplyr::select(-c(br_geom, dat_geom, grid_geom))
# write_csv(full_dat, "./data/02_data/2.3_dat.csv")
```

## GAM ARCHIVE
This model below makes sense and works in RUM. 
Things to check:
- does dem_bin includes all data 
-
```{r}
# dat <- dat %>%
#   mutate(Dem = ifelse(numYear < 2019, CaughtUndam, Dem)) %>%
#   mutate(Dem_bin = ifelse(Dem > 0, "1", "0"))

dat %<>% mutate(Dem_bin = as.factor(Dem_bin),
                launch2 = as.factor(launch2),
                side = as.factor(side),
                facYear = as.factor (facYear),
                lagoon = as.numeric(lagoon), 
                rug = as.numeric(rug), 
                depth = as.numeric(depth), 
                decDuration = as.numeric(decDuartion), 
                )

dat %<>% filter(depth < 600) # Lets see if this works 

binmod <- gam(Dem_bin ~ 
              s(lagoon, k = 3, bs = "cr") +
              s(rug, k = 3, bs = "cr") +
              s(depth, by = side, k = 3, bs = "cr") +
                launch2 +
                side+
              s(facYear, bs = 're'),
              # offset = decDuration, # test this with in teh moodel 
              method = "REML",
              family = binomial(),
              data = dat)
```






# BRTs
## BRT loop
```{r}
# define response
  resp <- which(colnames(dat) == "Dem") # poisson
  # resp <- which(colnames(dat) == "demph") # poisson
  # resp <- which(colnames(dat) == "dem_bin") # bernoulli

# defining covariates
  pred.vars <- c("Depth", "km_BR", "Pelagic", "Reef", "RReef", "Lagoon", "centroidLat", "decDuration") # covs for dem and dem_bin
  # pred.vars <- c("Depth", "km_BR", "Pelagic", "Reef", "RReef", "Lagoon", "centroidLat") # covs fro demph (remove duration)
    
# Set up CV
  k = 10 #number of folds
  p <- unique(dat$PersonID)
  p <- data.frame(PersonID = (sample(p)))# shuffle the data
  p$k <- c(rep(1:k, floor(nrow(p)/k)), sample(1:k, size = nrow(p) - length(rep(1:k, floor(nrow(p)/k)))))
  dat <- left_join(dat, p, by = "PersonID")
  table(dat$k) #k provides indicator for 10 fold cross validation

# define parameters for gbm.step to loop over
  step.loop <- expand.grid(lr = c(0.01, 0.001, 0.0001), # min
                           tc = c(5, 7, 10), # 1-5
                           bf = c(0.5, 0.65, 0.75), # max 0.75
                           ss = c(3, 5, 7)) 

  step.loop$n.trees <- NA ## add res.dev to grid
  step.loop$res.dev <- NA ## add res.dev to grid
  
  nmods <- nrow(step.loop)
  print(paste0("Making ", nmods, " models"))

for (i in 1:nrow(step.loop)) {
  print(i)
  
  tryCatch(
    {
      result <- gbm.step(data = dat,
                         gbm.x = pred.vars, ## covariates
                         gbm.y = resp, ## response
                         fold.vector = dat$k,
                         n.folds = k,
                         lr = step.loop[i, "lr"], ## ref to step.loop
                         tree.complexity = step.loop[i, "tc"], ## ref to step.loop
                         family = "poisson", ## distribution family
                         bag.fraction = step.loop[i, "bf"],
                         step.size = step.loop[i, "ss"],
                         plot.main = FALSE,
                         silent = T)
      
      step.loop[i, "nt"] <- result$n.trees
      step.loop[i, "res.dev"] <- result$self.statistics$mean.resid ### store res.dev in step.loop
      
      step.loop %<>% arrange(res.dev)
    },
    error = function(e) {
      # Handle the error if necessary
      print(paste("Error occurred in iteration", i, ":", e))
    }
  )
}
```
## Separate BRTs
```{r dp.step loop, warning=FALSE}
## Demersal catch
resp <- c("Dem") # poisson
pred.vars <- which(colnames(dat) %in% c("Depth", "km_BR", "centroidLat", "decDuration"))

dem_mod <- dismo::gbm.step(data = dat,
                           gbm.x = pred.vars, ## covariates
                           gbm.y = resp, ## response 
                           fold.vector = dat$k,
                           # offset = dat$decDuration,
                           n.folds = k,
                           lr = 0.0001, 
                           tree.complexity = 5, 
                           family = "poisson", ## distribution family
                           bag.fraction = 0.75,
                           step.size = 2
                         )

## demersal catch binary
resp <- which(colnames(dat) %in% "dem_bin") # bernoulli
pred.vars <- which(colnames(dat) %in% c("Depth", "km_BR", "centroidLat", "decDuration"))


dem_bin_mod <- dismo::gbm.step(data = dat,
                           gbm.x = pred.vars, ## covariates
                           gbm.y = resp, ## response 
                           family = "bernoulli",
                           tree.complexity = 5,
                           learning.rate = 0.01,
                           bag.fraction = 0.5,
                           step.size = 3,
                           fold.vector = dat$k,
                           n.folds = k,
                         )


## Demersal catch per hour
resp <- which(colnames(dat) == "demph") # poisson
pred.vars <- which(colnames(dat) %in% c("Depth", "km_BR", "centroidLat"))

dem_mph_mod <- dismo::gbm.step(data = dat,
                           gbm.x = pred.vars, ## covariates
                           gbm.y = resp, ## response 
                           fold.vector = dat$k,
                           n.folds = k,
                           lr = 0.01, 
                           tree.complexity = 3, 
                           family = "poisson", ## distribution family
                           bag.fraction = 0.5,
                           step.size = 5,
                           n.trees = 1000)
```


# ARCHIVE 
Explore differences between me and jons catch data 
```{r setup}
# looks like all zeros
# ggplot()+
#   geom_sf(data = full_dat$asc_geom) # full dat has complete grid
# 
# 
# ggplot()+
  # geom_sf(data = dat$asc_geom, aes(fill = as.factor(dat$CaughtUndam))) # some cels may be mission ig the fisher didnt do demersal fishing
  # geom_sf (data = full_dat$dat_geom)
# # 
# # # # summarise by geom to see spatial variation
# d <- dat %>% group_by(asc_geom) %>% summarise(caughtsum = sum(CaughtUndam))
# d <- dat %>% group_by(asc_geom) %>% summarise(caughtav = round(mean(CaughtUndam)))
# d <- dat %>% group_by(asc_geom) %>% summarise(caughtav = round(median(CaughtUndam)))
# # # 
# ggplot()+
#   geom_sf(data = d$asc_geom, aes(fill = as.factor(d$caughtsum)))
# # 
# ggplot()+
#   geom_sf(data = d$asc_geom, aes(fill = as.factor(d$caughtav)))
# 
# 
# # 
# # # catch is higher near boat ramps - somethings up 
# d <- n %>% filter(CaughtUndam > 0)
# d0 <- dat %>% filter(CaughtUndam == 0)
# # # 
# ggplot()+
# #   # geom_sf(data = d0$dat_geom, colour = "grey") +
#   geom_sf(data = d$dat_geom, aes(colour = d$CaughtUndam), alpha = 0.5)
# # 
# 
# 
# # # table(dat$Dem, dat$numYear)
# # table(dat$CaughtUndam, dat$numYear)
# # # table(dat$numYear)
# # # 
# j <- dat %>% filter(numYear < 2019)
# n <- dat %>% filter(numYear > 2019)
# # 
# # hist(j$CaughtUndam) # good
# hist(n$CaughtUndam) # good
# # hist(dat$CaughtUndam) # good
# 
# # summary(j$CaughtUndam) # median 10, max 65
# # table(j$CaughtUndam) # actually only 23 zeros hist looked ggo because of bins
# # # 
# summary(n$CaughtUndam) # median 6, max 37
# table(n$CaughtUndam) # more zeros but not zero inflated
# # 
# # hist(n$Dem) # looks good - outlier at 40
# # summary(n$Dem) # median 3, max 37 
# # table(n$Dem) # zero inflated - what i expect
# # 
# dat <- dat %>%
#   mutate(Dem2 = ifelse(numYear < 2019, CaughtUndam, Dem)) %>% # adding in jons dat to dem
#   mutate(Dem_bin2 = ifelse(Dem2 > 0, "1", "0")) %>%
#   mutate(caught_bin = ifelse(CaughtUndam > 0, "1", "0"))# binary
# # 
# hist(dat$Dem2)  # looks good
# summary(dat$Dem2) # median 8, max 65
# table(dat$Dem2) # 24 zeros, not inflated with uo to 20 people catch ing upto 8 fish
# # 
# # 
# ggplot()+
#   geom_bar(data = dat, aes(x = Dem_bin2)) # opposite to wha i would expect - data is telling me more people ctcah demersal sthan dont
  

# ggplot()+
#   geom_bar(data = dat, aes(x = caught_bin)) # same unexpected pattern
# 
# j %<>%  
#   mutate(caught_bin = ifelse(CaughtUndam > 0, "1", "0"))# 
# 
# ggplot()+
#   geom_bar(data = j, aes(x = caught_bin)) # same pattern again
# 
# 
# n %<>%  
#   mutate(caught_bin = ifelse(CaughtUndam > 0, "1", "0"))# 
# 
# ggplot()+
#   geom_bar(data = n, aes(x = caught_bin)) # same pattern again 
# 
# ggplot()+
#   geom_histogram(data = n, aes(x = CaughtUndam))
# 
# ggplot()+
#   geom_histogram(data = n, aes(x = Dem))
# 
# ggplot()+
#   geom_histogram(data = n, aes(x = Dem_bin))
# 
# ggplot()+
#   geom_histogram(data = dat, aes(x = CaughtUndam))
# 
# ggplot()+
#   geom_histogram(data = dat, aes(x = Dem))
# 
# ggplot()+
#   geom_sf(data = dat$asc_geom) +
#   geom_sf(data = n$dat_geom, aes(colour = n$Dem))
# 
# dat[, c("numYear", "CaughtUndam", "caught_bin", "Dem", "Dem_bin")]
```

there are alot of far away point like pt cloats which only have one observation where they caught 0 fish - this makes it look like you get lower catch the further away you are from the boat ramp, but really a smapling issue, your more likly to meet someone who has caught 10 fish from the poopular place than unpopular.

dat is modeling the data which was actually vistied by a recreator where choice = 1.
modelling fishers that have been targetting demersals (fishingtype)
we want to make a catch model which predicts the number of demersal fish caught across our observations

dependent will be dem or dem bin, explanatory variables will be spatial and fisher attributes


okay so jon data does not have whetehr they caught demersal/pelagic/nearshore fish etc. he just split his fishers into trolling and demersal So I will go back to modelling all people who targetted demersal na dall the fish they caught. Caught Undam or cuaght undam bin is the dependent var

I think this skews it alot because of all the nearshore fish so im going to have demersal catch and just add all of jon caught undam catch to dem  - this didnt help I still have less zeros than I would expect and catchs up beyond 60., Doing this only gives me 24/339 zeros 

investigate the difference between jon and my data - maybe mine is just really bad

another line of enquiry is the grid - i could make it smaller so the rugosity doesnt take over in the grid with the canyon.
